Develop HotFix Session추가

+++ J20180404/src/com/cheapmall/service/admin/UserAdminDeleteAction.java
@@ -7,6 +7,7 @@ import java.io.IOException;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 import com.cheapmall.dao.MemberDao;
 import com.cheapmall.service.CommandProcess;
@@ -20,6 +21,13 @@ public class UserAdminDeleteAction implements CommandProcess{
 		String check=request.getParameter("del");
 
 		try {
+			HttpSession session = request.getSession();
+			String auth =session.getAttribute("auth") == null ? null : session.getAttribute("auth").toString();
+			
+			if(auth == null) {
+				session.invalidate();
+				return "Admin.jsp";
+			}
 			int result=0;
 			MemberDao dao=MemberDao.getInstance();
 			result= dao.deleteUser(check);
